eShopAnalysis.ETL/eTLCartItemDW/requirements.txt
pandas
python-dotenv
pymongo
pyodbc
ipykernel
spacy
nltk